+++ src/message.js
@@ -159,7 +159,7 @@ class Message extends Syncable {
     if (options && options.fromServer) {
       this._populateFromServer(options.fromServer);
     } else {
-      this.sender = { userId: '', name: '' };
+      this.sender = { userId: '', name: '', displayName: '', avatarUrl: '' };
       this.sentAt = new Date();
     }
 
@@ -651,40 +651,40 @@ class Message extends Syncable {
    *
    * * layer.Constants.DELETION_MODE.ALL: This deletes the local copy immediately, and attempts to also
    *   delete the server's copy.
+   * * layer.Constants.DELETION_MODE.MY_DEVICES: Deletes this Message from all of my devices; no effect on other users.
    *
    * @method delete
-   * @param {number} deletionMode - layer.Constants.DELETION_MODE.ALL is only supported mode at this time
+   * @param {number} deletionMode
    */
   delete(mode) {
-    if (this.isDestroyed) {
-      throw new Error(LayerError.dictionary.isDestroyed);
+    if (this.isDestroyed) throw new Error(LayerError.dictionary.isDestroyed);
+
+    let queryStr;
+    switch (mode) {
+      case Constants.DELETION_MODE.ALL:
+      case true:
+        queryStr = 'mode=all_participants';
+        break;
+      case Constants.DELETION_MODE.MY_DEVICES:
+        queryStr = 'mode=my_devices';
+        break;
+      default:
+        throw new Error(LayerError.dictionary.deletionModeUnsupported);
     }
 
-    const modeValue = 'true';
-    if (mode === true) {
-      logger.warn('Calling Message.delete without a mode is deprecated');
-      mode = Constants.DELETION_MODE.ALL;
-    }
-    if (!mode || mode !== Constants.DELETION_MODE.ALL) {
-      throw new Error(LayerError.dictionary.deletionModeUnsupported);
-    }
-
-    if (this.syncState !== Constants.SYNC_STATE.NEW) {
-      const id = this.id;
-      const client = this.getClient();
-      this._xhr({
-        url: '?destroy=' + modeValue,
-        method: 'DELETE',
-      }, result => {
-        if (!result.success) Message.load(id, client);
-      });
-    }
+    const id = this.id;
+    const client = this.getClient();
+    this._xhr({
+      url: '?' + queryStr,
+      method: 'DELETE',
+    }, result => {
+      if (!result.success && (!result.data || result.data.id !== 'not_found')) Message.load(id, client);
+    });
 
     this._deleted();
     this.destroy();
-
-    return this;
   }
+
 
   /**
    * The Message has been deleted.
@@ -758,6 +758,8 @@ class Message extends Syncable {
     this.sender = {
       userId: message.sender.user_id || '',
       name: message.sender.name || '',
+      displayName: message.sender.display_name || '',
+      avatarUrl: message.sender.avatar_url || '',
     };
 
     this._setSynced();